+++ core/src/X11/window.c
@@ -150,10 +150,13 @@ static void xlib_window_destroy( sgui_window* this )
     if( TO_X11(this)->ic )
         XDestroyIC( TO_X11(this)->ic );
 
-    if( this->backend==SGUI_NATIVE )
-        sgui_canvas_destroy( this->ctx.canvas );
-    else
-        this->ctx.ctx->destroy( this->ctx.ctx );
+    if( this->ctx.canvas || this->ctx.ctx )
+    {
+        if( this->backend==SGUI_NATIVE )
+            sgui_canvas_destroy( this->ctx.canvas );
+        else
+            this->ctx.ctx->destroy( this->ctx.ctx );
+    }
 
     if( TO_X11(this)->wnd )
         XDestroyWindow( x11.dpy, TO_X11(this)->wnd );
